Extract `templateString()` method
